Диплом. Основной сервис.

+++ stats-server/service/src/main/java/ru/yandex/practicum/repository/HitRepository.java
@@ -3,6 +3,6 @@ package ru.yandex.practicum.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import ru.yandex.practicum.model.Hit;
 
-public interface HitRepository extends JpaRepository<Hit, Integer> {
+public interface HitRepository extends JpaRepository<Hit, Long> {
 
 }
